chore: Convert sheets on `/setting` route to use refs
